+++ src/main/scala/ot/dispatcher/plugins/externaldata/commands/WriteFile.scala
@@ -12,9 +12,24 @@ import ot.dispatcher.sdk.PluginUtils
  */
 class WriteFile(sq: SimpleQuery, utils: PluginUtils) extends ExternalFile(sq, utils) {
 
+  val mode: String = getKeyword("mode").getOrElse("overwrite")
+  val numPartitions: Option[String] = getKeyword("numPartitions")
+  val partitionBy: Option[String] = getKeyword("partitionBy")
+
   override def transform(_df: DataFrame): DataFrame = {
-    _df.write.format(format).mode(SaveMode.Overwrite).option("header", "true").save(absolutePath)
-    _df
+
+    val df = numPartitions match {
+      case Some(np) => _df.repartition(np.toInt)
+      case None => _df
+    }
+
+    val dfw = partitionBy match {
+      case Some(pb) => df.write.partitionBy(pb.split(",").map(_.trim):_*)
+      case None => df.write
+    }
+
+    dfw.format(format).mode(mode).option("header", header).save(absolutePath)
+    df
   }
 
 }