+++ src/main/scala/ot/dispatcher/plugins/externaldata/internals/ExternalFile.scala
@@ -12,12 +12,14 @@ import ot.dispatcher.sdk.{PluginCommand, PluginUtils}
  */
 class ExternalFile(sq: SimpleQuery, utils: PluginUtils) extends PluginCommand(sq, utils) {
 
-  val format: String = getKeyword("format").get
-  val path: String = getKeyword("path").get.replace("../","")
+  val format: String = getKeyword("format").getOrElse("parquet")
+  // TODO Secure path
+  val path: String = getKeyword("path").get
   val fs: String = pluginConfig.getString("storage.fs")
   val basePath: String = pluginConfig.getString("storage.path")
   val absolutePath: String = fs + new File(basePath,path).getAbsolutePath
-  log.info(s"Absolute path: $absolutePath. Format: $format")
+  val header: String = getKeyword("header").getOrElse("true")
+  log.info(s"Absolute path: $absolutePath. Format: $format. Header: $header")
 
   val requiredKeywords: Set[String] = Set("format", "path")
   val optionalKeywords: Set[String] = Set()