+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-01-12
+### Changed
+- Local resolver in build.sbt
+### Added
+- Arg "header" to readFile and writeFile commands.
+- Default value "parquet" to "format" arg in readFile and writeFile commands.
+- Arg "mode"  ('overwrite', 'append', 'ignore', 'error', 'errorifexists') to writeFile command.
+- Arg "numPartitions" to writeFile command.
+- Arg "partitionBy" to writeFile command.
+- Aliases "put" and "get" to writeFile and readFile commands.
+
 ## [1.0.3] - 2022-01-31
 ### Changed
 - plugin.conf.example.